Correct order for BIND() clauses
BIND()ing of variables needs to be done before using them, so this
change fixes the order of these statements.

This change is work-in-progress, there are more places to be checked.
Hint: search all templates in data/templates/*.html for "BIND(" and
ensure that variables are bound before thesy are used.